+++ packages/core-mobile/app/new/features/portfolio/assets/hooks/useTokenDetailFilterAndSort.ts
@@ -1,11 +1,5 @@
 import { IndexPath } from '@avalabs/k2-alpine'
 import { useCallback, useMemo, useState } from 'react'
-import {
-  TOKEN_DETAIL_FILTERS,
-  TOKEN_DETAIL_SORTS,
-  TokenDetailFilter,
-  TokenDetailSort
-} from 'store/balance'
 import { Transaction } from 'store/transaction'
 import { TransactionType } from '@avalabs/vm-module-types'
 import { sortUndefined } from 'common/utils/sortUndefined'
@@ -103,3 +97,34 @@ export const useTokenDetailFilterAndSort = ({
     data: filteredAndSorted
   }
 }
+
+// Token Detail
+export enum TokenDetailFilter {
+  All = 'All',
+  Sent = 'Sent',
+  Received = 'Received',
+  Bridge = 'Bridge',
+  Swap = 'Swap'
+}
+
+export enum TokenDetailSort {
+  NewToOld = 'Newest to oldest',
+  OldToNew = 'Oldest to newest'
+}
+
+type TokenDetailFilters = TokenDetailFilter[][]
+type TokenDetailSorts = TokenDetailSort[][]
+
+const TOKEN_DETAIL_FILTERS: TokenDetailFilters = [
+  [
+    TokenDetailFilter.All,
+    TokenDetailFilter.Sent,
+    TokenDetailFilter.Received,
+    TokenDetailFilter.Swap,
+    TokenDetailFilter.Bridge
+  ]
+]
+
+const TOKEN_DETAIL_SORTS: TokenDetailSorts = [
+  [TokenDetailSort.NewToOld, TokenDetailSort.OldToNew]
+]
